First version of network linearisation code.
Primitive networks that only consist of single input/output blocks can be sorted in such a way that all blocks get lined up in their correct execution order.
At the moment the sorting is done in a separate list.
After each sorting step, the wakeup() methods of the blocks that are connected to another block are called.